Finally got firebase stuff organized, able to protect routes that require auth
