+++ src/routes.js
@@ -1,5 +1,6 @@
 import Vue from 'vue'
 import VueRouter from 'vue-router'
+import firebase from './firebase'
 
 Vue.use(VueRouter)
 
@@ -17,10 +18,10 @@ const routes = [
   { path: '/components', name: 'components', component: Components },
   { path: '/login', name: 'login', component: Login},
   { path: '/register', name: 'register', component: Register},
-  { path: '/dashboard', name: 'dashboard', component: Dashboard},
   { path: '/roadmap', name: 'roadmap', component: Roadmap},
 
-  { path: '/new-character', name: 'new-character', component: NewCharacter}
+  { path: '/dashboard', name: 'dashboard', component: Dashboard, meta: {requiresAuth: true}},
+  { path: '/new-character', name: 'new-character', component: NewCharacter, meta: {requiresAuth: true}}
 ]
 
 const router = new VueRouter({
@@ -29,6 +30,15 @@ const router = new VueRouter({
     return { x: 0, y: 0 };
   },
   routes
+})
+
+router.beforeEach(async (to, from, next) => {
+  const requiresAuth = to.matched.some(record => record.meta.requiresAuth);
+  if (requiresAuth && !await firebase.getCurrentUser()) {
+    next('login')
+  }else {
+    next()
+  }
 })
 
 export default router;
