+++ src/routes/NewCharacter.vue
@@ -30,8 +30,8 @@
 
 <script>
 import { mapGetters } from 'vuex' 
-import { db } from '../main'
-import firebase from 'firebase'
+import { db } from '../firebase'
+import firebase from '../firebase'
 
 import Title from '@/components/Title.vue'
 import Input from '@/components/Input.vue'
